Issue 391/improve and refactor the crypto story in rama new boring config (#560)
Goal of this PR: Move away from ClientConfig defined in rama-net/tls for
Boring Tls config (part of https://github.com/plabayo/rama/issues/391)

Things done to achieve that:
- TlsAcceptorBuilder
  - We pass this builder around so we can do modifications where needed
- By using "base builders" together with getters for all fields we can
chain configs in a cheap way, and it is also easy to do. These base
builders are inside an Arc so cloning them is very cheap
- We only Build acceptor data when we actually need it. This allows use
to build the Boring config by using almost only references vs cloning
everything like we did before, and allows use to modify and chains these
builders as needed
- rama-tls-boring now imports rama-ua when ua emulation is enabled and
will use TlsProfile to create a builder with everything needed to apply
this profile.
- This is a new service/layers to clearly indicate that this is
something extra and this service will get more complex overtime
- This service will append the builder needed to the one currently in
context or create a new one if needed
- This service also has support for global overwrites eg if you want to
emulate but always want to force http1 for example. It does this by
appending the overwrites to the builder chain
- Extend generate_set_and_with to also support Result types. We can make
this macro "much smarter", but for now I prefer the copy paste as it
makes it very easy to read and understand

Things not part of this PR
- Completely refactor EasyClient to make use of this. This will be done
in https://github.com/issues/assigned?issue=plabayo%7Crama%7C529
- Change what we store in TlsProfile. For now we leave the ClientConfig
in there, but this will be changed in the future once we have a better
idea of what to store instead