Use `instrument` instead of `Span::enter` in async
The documentation for `tracing::Span` mentions that it's better not to
hold a span's guard across await points, because an unrelated task might
end up executing while remaining in the entered span.

Instead, the docs suggest using `Instrument::instrument`, which ensures
the span is entered and exited every time the future is polled.