add datastar as builtin support (#599)
Move from rama support in datastar crate to having datastar directly
working in rama

Problem with the sdk crate of datastar is that it didn't add much to
begin with once you have built-in SSE support.
Next to that there is not a lot you can improve in terms of allocations
and type safety given other frameworks their
SSE support doesn't help us here at all given that datastar events is
multiline with their own protocol.

- [x] initial implementation
- [x] working hello world example
- [x] unit tests read-write (individual ones as well as enum)
- [x] integration test hello-world example
- [x] update docs
- ~~extra example: adanvaced~~