Add acme client support (#603)
What this PR does
- Rama Acme crate
  - Proto folder which contains all the needed types
- Client implementation with all the needed logic to request certificate
- Some improvements in rama-crypto
- Add pki_types and x509 as dependencies so we can start centralizing
these deps
  - Implement Signer for EcdsaKey so it can be used direct for JWS
  - Make all JWK fields public (should have already been the case)
- Add constants for NO_PAYLOAD, EMPTY_PAYLOAD to avoid common pitfalls
here
  - Support missing fields in JWS and JWSFlattened
- JWS Verify trait can return output now. In some cases the Verifier
will decode some data that the application will also need later. Without
support to return this the app would otherwise have to decode again (use
case will be in acme server where things like JWK are needed to verify
signature but also for account management)
- Store decode signature in JWS decoded (makes more sense and is easier
to work with)
- Examples for acme using peble to verify them. Http challenge, tls
challenge using rustls and boring. They currently all use the same port
since we dont have testing for them yet
- Expose some getters and modify some visibility in rama headers to make
them more useful on server side

What it doesn't do. Some of these will be implement very soon in
followup PRs or when we need them. If needed some items from this list
can still be included in this PR
- Full ACME client, things like account management, revoking
certificates
- Acme server (other issue for that one)
- Tests for examples: they depend on rama acme server implementation
- Run pebble tests with justfile or in cicd. Quite hard to get
everything working in docker with dns and stuff. Currently searching for
ways but dont think it has to block this PR?
- Examples have certificates at the end but we don't really use them.
Not sure if that is needed?
- Make it easier to integrate acme client setup in common service stacks
- Dynamic acme client that updates certificates on the fly
- Lots of small improvements probably after battle testing this live