Issue 498/support auto proxy connector http+https+socks5+socks5h (#659)
- Adds socks5(h) support to EasyWebClient
- Add proxy connector that supports both http and socks5 proxies and
delegates to a specific connector depending on proxy protocol
- Kept this connector specific (read: not generic) to our
HttpProxyConnector/Socks5ProxyConnector since this is used for
EasyWebClientBuilder specifically and saw no real value in making this
generic. For special use cases we already support bringing a custom
proxy connector
- Another option was also. Currently you can configure existing
ProxyConnectors in `optional` mode which forwards traffic to inner one
if proxy is not configured. I could have modified this to also forward
traffic to inner svc if Proxy protocol was wrong instead of returning
error like we do now, but this felt ugly and error prone in the long run
- Right now I focussed on http(s) + socks5(h) only, not a generic amount
of proxy connectors. Once we add one more I have some "scalable" ideas
for this but didn't want to overcomplicate the current setup for
something we might not support for quite some time
- Move EasyWebClient to rama crate as now its not specific to http
anymore
- Move autobahn example to rama examples so it can use EasyWebClient