Issue 462/revisit service api (core) design in 0.3 milestone# removing state from context (#685)
Part of #462, this PR removes State from Context. Where state is still
needed it has been moved to `ctx.extensions` and in most cases been
added with `AddExtensionLayer`.

Logic for with_state has also been removed in favor of
`AddExtensionLayer`. Once we have a solution for fully typed (localized)
state again we could probably add this again, but until then it has been
removed so we can start with a clean slate.