allow wildcard domains (rama-net)
NOTE: not all domains that parse with wildcard suffix
are also acceptable in wild. E.g. we allow '*.com' as a valid
wildcard domain while nothing in the wild of meaning (e.g. LetsEncrypt)
would ever accept such a wildcard domain to gen certs for...

It is still up to you as the user of these apis to check you are really
using valid/meaningful domains

If you know a const-way to easily use something alike to psl
to validate wildcard domains on public suffixes, do submit a patch