Issue 462/revisit service api (core) design in 0.3 milestone#2 moving extensions (#706)
What this PR does (for #462):
- Add extensions to all Request/Response/Connection types
- Move ctx extensions to these extensions instead and remove extensions
from context
- ExtensionsRef, and ExtensionsMut trait so we can implement this
generically
- `debug-extensions` feature that implement a debug impl for extensions
that can print all type_names stored in extensions, this made debugging
extensions issues a lot easier
- ChainableExtensions -> this is something i want to improve (see next
points) but for now it does help already
- Extractor to get extensions, currently this clones them, but we can
probably do this similar to how we consume body. To be checked in
followup pr also
- Add ExtensionsMut + Send bound on ConnectorService's connection and
Send + 'static bounds on Error. This is similar to how we set bounds on
Service trait and makes usage easier all around

What this Pr doesn't do
- Removing ctx. Some very minor things need to be migrated, but removing
ctx from all API has a pretty big impact and I want to combine this with
thinking how we expose/make these APIs as I believe we can do better.
This is partially related to the never caching remark and making traits
for that
- Huge refactors to be consistent. This will happen in follow up PR so
we can focus specifically on that instead of getting lost in this big
PR. Some refactoring already happened where needed. This includes
renaming things, making this private vs public...
- A better design to transfer extensions. Right now we always transfer
all of them, but depending on the place/service this can be very tricky
and error prone (especially around upgrades). I'm already playing with
some ideas to make this much nicer, but will implement this in a follow
up PR. In general I believe we can make extensions easier and less error
prone to work with

Unanswered questions that will require some input / more playing time:
- Right now we are able to add extensions to pretty much all connection
types except the inner http-core types. Adding and passing them all
through the stack there is pretty complex, but it might be worth it down
the line. If extensions are everywhere it's pretty transparent how they
are passed, right now we need to remember on which levels to pass them
and where they were skipped. On the other hand, not having to think
about extensions in places makes the code easier there and more
focussed. Curious about what you think @GlenDC , but either way I still
believe we can do much better in general, so might be a question we
answer once all the other improvements are done