Issue 462/revisit service api (core) design in 0.3 milestone#3 no remove from extensions (#715)
- No more support for `remove` on extensions, this makes them much more
predictable and will leave a trace of info. Especially once we add
support for "parent" extensions
- We can remove `clear_extensions_safely`. The underlying original logic
in hyper was needed to prevent a deadlock that got triggered because
extensions were dropped that included a lock in the drop, while holding
that same lock in these functions. Because we now always keep extensions
around we are not affected by that problem
- Refactor upgrade::on to upgrade::handle_upgrade where we never have
support None upgrades. Instead we throw error if we did an unexpected
upgrade, or we already handled the upgrade
- Remove legacy tls logic that should have been removed already
- Add parent extensions (this includes the todo from setting request
extensions as parent extensions for response)

For separate PR:
- On server side we use RamaHttpService to store extensions from io, and
then reapply them. Should be pretty doable to do this in http-core where
we have io access and create the actual request. Should also make this
less error prone. But I has to potential to be a quite invasive change,
so wont do that here