Issue 462/revisit service api (core) design in 0.3 milestone#4 state for http endpoints (#720)
- Add state for `WebService` and `Router` by adding
`IntoEndpointServiceWithState`. And providing this to our extractors
- Add `FromRef` trait and macro to auto derive it
- Refactor FactoryTrait `EndpointServiceFn` with pattern T for any
Function to use nested tuples so we have predictable signatures. Will
add a paragraph to patterns chapter for this soon
- Refactor `Arc<BoxService>` to `BoxService` since this is now uses
`Arc` internally.
- Update examples to use this state

For separate PRs:
- Docs about pattern used in `EndpointServiceFn` to make predictable
generics for function signatures
- State for service_fn. Once this is ready also I will expand the docs
around state and explain all the different option for each place and the
pros and cons

part of #462