extensions inside rama-http-core (#727)
Extensions are now supported throughout rama-http-core, for both h1 and
h2. Because of this we can now remove a lot of extension hacks and
properly set/pass extensions where we create requests/responses.

Some TODOs have been added about performance, and I have some more in
mind (related to breadcrumbs efforts. Will create a ticket for this
soon, but I see that as part of rama 0.4 efforts, as without benchmarks
and extensive testing it's hard to tell the impact of these changes.

part of #462