Remove request inspectors (#750)
This fits in the effort of #462 because it removes extra complexity and
goes back to the core idea of everything is a service.
The main benefit here is removing extra complexity while still keeping
the same functionality (simpler = better most of the time).

What is done in this PR
- Remove request inspector concept/code
- Refactor request inspector service to normal services/layers
- Remove http_req_inspector in http connector. While this is a useful
thing, it should not be part of the core http connector
- Refactor EasyHttpClient to be more explicit: build_connector,
connector_builder... vs build (what do we build?)
- Support jit_layers in EasyHttpClient, here this concept is still
needed, but since this is a higher level abstraction it makes sense here

---------

Co-authored-by: glendc <glen@plabayo.tech>