Issue 756/Rename AddExtension to AddInputExtension and add AddOutputExtension (#761)
Couple of thoughts I had while doing this one:
* I am not sure if my example at the top of `add_extension` with the
`ResponseCounter` is a good one, happy to change it/remove it if we
don't need an example
* The `serve` function for `AddInputExtension` used to return a
`Future<>` but for `AddOutputExtension` I needed to await to get the
response, so I made both of them consistent and converted them into
`async fn`s. This is also consistent with `GetInputExtension` and
`GetOutputExtension`. I am not super familiar with async rust (let alone
rust) so I am not 100% sure about the implications of this change but I
have done a bit of looking into it so `async fn` is the same as
returning a `Future<>` from a function that wraps the original function
body in an `async move`. The things im not sure about are:
* `'_` - im not sure we need this, presumably we want the returned value
to live as long as the args, is that the default for async functions?
* `Send` - not sure if we need to enforce it on the return here, `Input`
has the trait enforced and its the only one we take a mutable reference
to

Relates to #759