Patch/fix proxy compression (#763)
If you currently try to MITM proxy compressed chunked encoding and
decompress on the fly you'll notice unexpected EOF errors... This patch
fixes that.

While I'm at it I'm also removing compress adepter as it honestly is
overkill for something that is much better solved with a separate
compressor and decompressor. This way we keep it also a lot simpler and
keep these each on their own side of the river.